+++ lib/ngcc/index.ts
@@ -1,0 +1,7 @@
+export * from './symbol';
+export * from './directive-symbol';
+export * from './module-symbol';
+export * from './pipe-symbol';
+export * from './project-symbols';
+export * from './provider-symbol';
+export * from './resource-resolver';
