+++ lib/ngcc/module-symbol.ts
