+++ lib/ngcc/pipe-symbol.ts
