+++ lib/ngcc/project-symbols.ts
