+++ lib/ngcc/resource-resolver.ts
