+++ lib/ngcc/symbol.ts
