+++ test/ngcc/dep-parsing.spec.ts
@@ -1,5 +1,6 @@
-import { ProjectSymbols } from '../';
-import { resourceResolver } from './utils/resource-resolver';
+import { ProjectSymbols } from '../../lib/ngcc';
+import { resourceResolver } from '../utils/resource-resolver';
+import { getConfig } from './get-tsconfig';
 
 const defaultErrorReporter = (e: any, path: string) => console.error(e, path);
 
@@ -8,7 +9,7 @@ describe('DirectiveSymbol', () => {
     let program: string;
 
     beforeEach(() => {
-      program = __dirname + '/../../test/fixture/deps/tsconfig.json';
+      program = getConfig('deps');
     });
 
     it(`should parse dependencies`, () => {
