+++ test/ngcc/get-tsconfig.ts
@@ -1,0 +1,5 @@
+import { join } from "path";
+
+export function getConfig(name: string) {
+  return join(__dirname, '/../../../test/fixture', name, 'tsconfig.json');
+}
