+++ test/ngcc/provider-symbol.spec.ts
@@ -1,5 +1,6 @@
-import { ProjectSymbols } from '../';
-import { resourceResolver } from './utils/resource-resolver';
+import { ProjectSymbols } from '../../lib/ngcc';
+import { resourceResolver } from '../utils/resource-resolver';
+import { getConfig } from './get-tsconfig';
 
 const defaultErrorReporter = (e: any, path: string) => console.error(e, path);
 
@@ -8,7 +9,7 @@ describe('ProviderSymbol', () => {
 
   describe('basic example', () => {
     beforeEach(() => {
-      program = __dirname + '/../../test/fixture/routing/tsconfig.json';
+      program = getConfig('routing');
     });
 
     it(`should provide access to the module's providers`, () => {
@@ -35,7 +36,7 @@ describe('ProviderSymbol', () => {
 
   describe('basic primitive token example', () => {
     beforeEach(() => {
-      program = __dirname + '/../../test/fixture/basic-primitive-token/tsconfig.json';
+      program = getConfig('basic-primitive-token');
     });
 
     it('should discover transitive dependencies', () => {
